+++ lib/celery_script/commands/_if.ex
@@ -1,0 +1,71 @@
+defmodule Farmbot.CeleryScript.Command.If do
+  @moduledoc """
+    If
+  """
+
+  alias Farmbot.CeleryScript.Command
+  alias Farmbot.CeleryScript.Ast
+  require Logger
+
+  @behaviour Command
+
+  @doc ~s"""
+    Conditionally does something
+      args: %{_else: Ast.t
+              _then: Ast.t,
+              lhs: String.t,
+              op: "<" | ">" | "is" | "not",
+              rhs: integer},
+      body: []
+  """
+  @spec run(%{}, []) :: no_return
+  def run(%{_else: else_, _then: then_, lhs: lhs, op: op, rhs: rhs }, []) do
+    lhs
+    |> eval_lhs
+    |> eval_if(op, rhs, then_, else_)
+  end
+
+  # figure out what the user wanted
+  @spec eval_lhs(String.t) :: integer | {:error, String.t}
+
+  defp eval_lhs("pin" <> num) do
+    num
+    |> String.to_integer
+    |> Farmbot.BotState.get_pin || {:error, "pin" <> num}
+  end
+
+  defp eval_lhs(axis) do
+    [x, y, z] = Farmbot.BotState.get_current_pos
+    case axis do
+      "x" -> x
+      "y" -> y
+      "z" -> z
+      _ -> {:error, axis} # its not an axis.
+    end
+  end
+
+  @spec eval_if({:error, String.t} | integer, String.t, integer, Ast.t, Ast.t)
+    :: no_return
+
+  defp eval_if({:error, lhs}, _op, _rhs, _then, _else) do
+    Logger.info "Could not evaluate left hand side: #{lhs}"
+  end
+
+  defp eval_if(lhs, ">", rhs, then_, else_) do
+    if lhs > rhs, do: Command.do_command(then_), else: Command.do_command(else_)
+  end
+
+  defp eval_if(lhs, "<", rhs, then_, else_) do
+    if lhs < rhs, do: Command.do_command(then_), else: Command.do_command(else_)
+  end
+
+  defp eval_if(lhs, "is", rhs, then_, else_) do
+    if lhs == rhs, do: Command.do_command(then_), else: Command.do_command(else_)
+  end
+
+  defp eval_if(lhs, "not", rhs, then_, else_) do
+    if lhs != rhs, do: Command.do_command(then_), else: Command.do_command(else_)
+  end
+
+  defp eval_if(_, _, _, _, _), do: Logger.info "bad if operator"
+end
