+++ lib/celery_script/commands/calibrate.ex
@@ -1,0 +1,24 @@
+defmodule Farmbot.CeleryScript.Command.Calibrate do
+  @moduledoc """
+    Calibrate
+  """
+
+  alias Farmbot.CeleryScript.Command
+  alias Farmbot.Serial.Handler, as: UartHan
+  @behaviour Command
+
+  @doc ~s"""
+    calibrates an axis:
+      args: %{axis: "x" | "y" | "z"}
+      body: []
+  """
+  @spec run(%{axis: String.t}, []) :: no_return
+  @lint {Credo.Check.Refactor.PipeChainStart, false}
+  def run(%{axis: axis}, []) do
+    case axis do
+      "x" -> "F14"
+      "y" -> "F15"
+      "z" -> "F16"
+    end |> UartHan.write
+  end
+end
